# server/docs/swagger.yaml
definitions:
  dao.MemosData:
    properties:
      content:
        type: string
    type: object
  dao.MoodReader:
    properties:
      highlights:
        items:
          $ref: '#/definitions/dao.MoodReaderHighlights'
        type: array
    type: object
  dao.MoodReaderHighlights:
    properties:
      author:
        type: string
      note:
        type: string
      text:
        type: string
      title:
        type: string
    type: object
  dao.UrlStruct:
    properties:
      url:
        type: string
    type: object
  jwt.User:
    properties:
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    type: object
  main.GeneralAllStruct:
    properties:
      content:
        type: string
      file_key:
        default: test.md
        type: string
      mod:
        enum:
        - append
        - cover
        type: string
    type: object
  main.ObV1ModMdText:
    properties:
      content:
        type: string
      day:
        type: string
      line_num:
        type: integer
      old:
        type: string
    type: object
  main.SendMailStruct:
    properties:
      content:
        type: string
      subject:
        type: string
    type: object
  main.SimpReadWebHookStruct:
    properties:
      content:
        type: string
      desc:
        type: string
      note:
        type: string
      tags:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
  main.WeChatInfoStruct:
    properties:
      content:
        type: string
    type: object
  talk.TalkStruct:
    properties:
      content:
        type: string
    type: object
externalDocs:
  url: https://kkbt.gitee.io/obcsapi-go/
info:
  contact:
    url: https://gitee.com/kkbt/obcsapi-go
  description: 基于 Obsidian S3 存储， CouchDb ，本地存储和 WebDAV 的后端 API ,可借助 Obsidian 插件 Remotely-Save
    插件，或者 Self-hosted LiveSync (ex:Obsidian-livesync) 插件 CouchDb 方式同步，保存消息到 Obsidian
    库。该调试页面大部分仅提供对 Headers-Token 验证方式的支持，其他如 Query-token，Headers-Authorization 除了特殊的几个其他并不支持。可以使用
    https://hoppscotch.io/ 或者 Postman 之类的工具，或者使用 VsCode 插件 REST Client ，使用 REST Client
    可以在 https://gitee.com/kkbt/obcsapi-go/tree/master/http ，找到测试文件
  title: Obcsapi
  version: v4.2.5 版本
paths:
  /api/sendmail:
    post:
      consumes:
      - application/json
      parameters:
      - description: SendMailStruct
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/main.SendMailStruct'
      produces:
      - application/json
      responses: {}
      security:
      - Token: []
      summary: 邮件通知
      tags:
      - 通知
  /api/v1/cacheupdate:
    post:
      parameters:
      - description: 更新文件 FileKey 完整的
        in: query
        name: key
        required: true
        type: string
      responses: {}
      security:
      - JWT: []
      summary: 更新文件的缓存
      tags:
      - 前端
  /api/v1/daily:
    get:
      description: 默认一周以前的查找缓存返回 即 <= -7 且不允许请求 一年之前的日记
      parameters:
      - description: 请求几天前的
        in: query
        name: day
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - JWT: []
      summary: Memos 请求
      tags:
      - 前端
  /api/v1/daily/nocache:
    get:
      parameters:
      - description: 请求几天前的
        in: query
        name: day
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - JWT: []
      summary: Memos 请求 (无缓存)
      tags:
      - 前端
  /api/v1/line:
    post:
      consumes:
      - application/json
      description: 根据行号修改内容，line_num 大于原文件行数，如 9999 新增 Memos 。需要原文件不完整 FileKey 和原来的行的内容进行校验。成功后返回更新后的内容。
      parameters:
      - description: 根据行号修改内容
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/main.ObV1ModMdText'
      produces:
      - application/json
      responses: {}
      security:
      - JWT: []
      summary: 根据行号修改内容
      tags:
      - 前端
  /api/v1/random:
    get:
      responses: {}
      security:
      - JWT: []
      summary: 随机回顾
      tags:
      - 前端
  /api/v1/sayHello:
    get:
      consumes:
      - text/plain
      - application/octet-stream
      produces:
      - application/json
      responses: {}
      security:
      - JWT: []
      summary: JWT 测试接口
      tags:
      - 前端
  /api/v1/talk:
    post:
      consumes:
      - application/json
      parameters:
      - description: TalkStruct
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/talk.TalkStruct'
      produces:
      - text/plain
      responses: {}
      security:
      - JWT: []
      summary: 指令模式接口
      tags:
      - 前端
  /api/wechatmpmsg:
    post:
      consumes:
      - application/json
      parameters:
      - description: WeChatInfoStruct
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/main.WeChatInfoStruct'
      produces:
      - application/json
      responses: {}
      security:
      - Token: []
      summary: 微信通知
      tags:
      - 通知
  /info:
    get:
      produces:
      - application/json
      responses: {}
      summary: 服务器信息与测试接口
  /login:
    post:
      consumes:
      - application/json
      description: 前端登录
      parameters:
      - description: User
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/jwt.User'
      produces:
      - application/json
      responses: {}
      summary: 前端登录
      tags:
      - 前端
  /ob/fv:
    post:
      consumes:
      - text/plain
      - application/octet-stream
      description: 安卓软件 fv 悬浮球使用的 API 用于自定义任务的 图片、文字
      parameters:
      - description: fv payload 內容
        in: body
        name: 內容
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses: {}
      security:
      - Token: []
      summary: fv 悬浮球使用的 API
      tags:
      - Ob
  /ob/general:
    post:
      consumes:
      - application/json
      description: 通用 API 接口,添加 Memos
      parameters:
      - description: MemosData
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/dao.MemosData'
      produces:
      - application/json
      responses: {}
      security:
      - Token: []
      summary: 通用 API 接口 Memos
      tags:
      - Ob
  /ob/general/{token}:
    post:
      consumes:
      - application/json
      description: 通用 API 接口,添加 Memos
      parameters:
      - description: 设定的 token 值
        in: path
        name: token
        required: true
        type: string
      - description: MemosData
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/dao.MemosData'
      produces:
      - application/json
      responses: {}
      summary: 通用 API 接口 (Memos Flomo Like API)
      tags:
      - Ob
  /ob/generalall:
    get:
      description: 通用 API 接口，获取所有文件。需要配置声明允许使用该接口
      parameters:
      - description: 文件名，有路径，如 dir/text.md
        in: query
        name: filekey
        required: true
        type: string
      produces:
      - text/plain
      responses: {}
      security:
      - Token: []
      summary: 通用 API 接口 All
      tags:
      - Ob
    post:
      consumes:
      - application/json
      description: 通用 API 接口，覆盖修改或增添所有文件。需要配置声明允许使用该接口
      parameters:
      - description: GeneralAllStruct
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/main.GeneralAllStruct'
      produces:
      - application/json
      responses: {}
      security:
      - Token: []
      summary: 通用 API 接口 All
      tags:
      - Ob
  /ob/moonreader:
    post:
      consumes:
      - application/json
      description: 静读天下使用的 API，标注-设置-ReadWise 设置该路径和 token 值即可
      parameters:
      - description: MoodReader
        in: body
        name: 划线和标注
        required: true
        schema:
          $ref: '#/definitions/dao.MoodReader'
      produces:
      - application/json
      responses: {}
      security:
      - AuthorizationToken: []
      summary: 静读天下使用的 API
      tags:
      - Ob
  /ob/sr/webhook:
    post:
      consumes:
      - application/json
      description: SimpRead 简悦 WebHook POST 简悦 WebHook 保存文章
      parameters:
      - description: SimpRead 简悦 POST
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/main.SimpReadWebHookStruct'
      responses: {}
      security:
      - Token: []
      summary: 简悦 WebHook 保存文章
      tags:
      - Ob
  /ob/url:
    post:
      consumes:
      - application/json
      description: 裁剪网页
      parameters:
      - description: MemosData
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/dao.UrlStruct'
      produces:
      - application/json
      responses: {}
      security:
      - Token: []
      summary: 裁剪网页
      tags:
      - Ob
securityDefinitions:
  AuthorizationToken:
    in: header
    name: Authorization
    type: apiKey
  JWT:
    in: header
    name: Authorization
    type: apiKey
  Token:
    in: header
    name: Token
    type: apiKey
swagger: "2.0"